requirements.txt
asgiref==3.7.2
Django==5.0
django-admin-interface==0.27.0
django-colorfield==0.11.0
django-jazzmin==2.6.0
django-jet-reboot==1.3.5
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
django-import-export

